requirements.txt
invoke==1.6.0
requests==2.26.0
markdown==3.3.4
Pillow==8.4.0
psycopg2-binary==2.9.2
Werkzeug==2.0.2

# django
Django==3.2.8
django-environ==0.7.0
django-extensions==3.1.3
djangorestframework==3.12.4
django-cors-headers==3.10.0

# Celery and flower
celery==5.2.0
django-celery-beat==2.2.1
flower==1.0.0

# Test
pytest==6.2.5
pytest-django==4.4.0
factory-boy==3.2.0
pytest-factoryboy==2.1.0

# code quality
black==21.10b0